m_lamp_oil6.blend
I created a 3D flame model in a separate file by applying a displace modifier to a magic texture, then extruding upwards along the Z axis to create a flame shape. Then I copied and pasted the model into my Lantern Oil file and placed it on top of the cork as a decoration.