m_lamp_oil7.blend
In this step, I shade smoothed the Lamp Oil bottle and added a coloured texture to both the bottle and the flame. For the texture on the bottle, I added the glass texture to the orange base colour. I then added a gradient texture. I came to the conclusion that the texture is way too shiny and that the gradient it vertical instead of horizontal. In future iterations I will make alterations to fix this problem.